Isolate setupData by unmarshalling it each time
This removes the ability to change setupData between calls of the
default function. This removes a race condition where multiple VUs are
changin the setupData at the same time.
This change also makes certain that each call to the default function and
teardown get a setupData that has the same contents making it less
confusing if you change the setupData inside any of them for some reason.