Report threshold failure correctly to output cloud fix #849
Previously if threshold fails once it marked as failed forever. This is
okay for local run because what we actually report at the end is if the
last threshold check has failed so this threshold failed marking is not
taken into consideration. But the cloud collector and the cloud
collector only takes this as the final result.

This probably should be fixed in some better way such that the final
threshold result is actually given to the collectors to report (if they
do) but this is outside the scope of this commit.

The fix is simply to mark the threshold as failed every time it is
checked this way if it fails once and then it stops failing it will
be marked as not failed again.