lib: better error message for timeout error
TimeoutError is currently very uninformative, even if there are
duplicated error messages.

This PR change TimeoutError to a struct, which contains both the place
and the timeout duration, so it helps constructing error message with
more useful information to user.

While at it, also prevent duplicating timeout error message for setup
and teardown.

Close #702