cmd: use default config dir instead of panic
In #1162 use new os.UserConfigDir to get user config directory, and it will
panic when encounter an error. This is a backward in-compatible behavior and
can lead to bad UX.

This PR restore previous #1162 behavior. Instead of panic, we return a
default value for config dir if we can figure out user config directory.

Close #1176