lib: use vu ctx instead of request ctx for pushing metrics
PushIfNotDone is intended to use with VU context instead of request
context. This is a bug in current code base, but it is not triggered due
to the bug in Go net/http. A request context can be in cancelled state
due to be cancelled or timeouted. PushIfNotDone won't see the difference
and will skip pushing metrics for timeouted requests.

Fixing it by passing VU context to PushIfNotDone instead of request ctx.

Fixes #1260