Add responsive progress bars
Progress bars will now dynamically resize depending on terminal window
width, being replaced with percentages if "squished" too far.

It works best on *nix platforms because of SIGWINCH, while on Windows
retrieving the size is required before each render iteration, and isn't
as responsive.

Terminal multiplexers are a bigger issue: tmux seems to buffer
SIGWINCH/terminfo and is plain broken on versions including the current
latest stable (3.0a) (see issue #2005[1]). This was fixed on `master`,
but is yet to be released, and even with the fix the experience is far
from ideal (lags quite a bit, no continuous resize). GNU screen also has
some rendering issues I haven't looked too much into yet.

Untested on macOS terminals, but I expect it to work as well as on Linux.

Terminals tested:
- Linux: st, urxvt, xterm
- Windows: cmd.exe and PowerShell standalone, and both inside the new
  Windows Terminal[2] app. Resizing works fine in all cases, though
  isn't as responsive as on Linux.

Part of #1279

[1]: https://github.com/tmux/tmux/issues/2005
[2]: https://github.com/microsoft/terminal