Fix TestVURunInterrupt by creating the context just before it's used
Previously what was happening that the context could timeout by the time
RunOnce is called which will lead to it seeing that and exiting with the
context error (timeout in this case).

This is easily fixable by creating the context just before it is used.
Which also makes it possible to decrease the context timeout
significantly, as it now doesn't need to last newVU for example.