merge steps in VLV executors for more stability (#1496)
Also fix VLV not respecting gracefulStop since 270fd91c, add test for it
and try to make some other VLV tests more stable.

The original code (pre 270fd91c) did go through all the raw steps
(ramp up/down) and the graceful stop ones concurrently and when the raw
steps ended it will return from Run, but will start a goroutine to go
through the remainign graceful stop ones. The most important of which (
at least in my current understanding) being the last one which is the
actuall gracefulStop end which will make all VUs stop. The reason why
this actually worked is that it also waiting on all activeVUs to have
ended being active, being returned in the new terminology, before it
actually cancels the context for all of them. So if the VUs manage to
end iterations in this time the executor will end earlier if not the
gracefuStop will make them.

270fd91c broke the returning of VUs which lead to the executor never
returning and moving the cancel "before" the waiting for activeVUs fixed
that at the cost of gracefulStop not being taken into account, but with
no tests, nobody noticed.

Here I basically revert that especially because vuHandle now returns VUs
when stopped.

fixes #1473

Co-authored-by: na-- <n@andreev.sh>