Switch to awscli, trigger CloudFront invalidation
The --cf-invalidate option doesn't seem to work for s3cmd, and its
documentation is lacking, so we might as well switch to the official
awscli tool, which is a good idea for future maintenance regardless.

This requires the cloudfront:CreateInvalidation permission to be allowed
for the specific CF distribution in the IAM policy of the user.