Explicitly disable GPG check for RPM repo metadata
After some investigation, we decided to disable this explicitly to avoid
it failing for users who have `repo_gpgcheck` enabled globally.

Our RPM packages are signed, but we're not planning to enable this check for
repo metadata since it's known to cause issues[1,2] and it's not enabled
even for main Fedora repositories.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1768206
[2]: https://bugzilla.redhat.com/show_bug.cgi?id=1636743