Better catch SyntaxError when babel returns wrong error type
for some reason Babel does not return SyntaxError in some cases where it
should. In a lot of those cases goja is *correctly* catching the syntax
error but we then proceed to try to run the code through babel.

This might be better of knowing that it expects a SyntaxError, but given
that this won't be what the end user gets (or any error) I went with
requiring that both babel and goja error out and overwrite babel type in
case both reported an error.