Panic on RegisterCallback result being used more than once
This should help with actually finding problems with code using it.

I have already seen multiple times code which seems okay but in fact
will call the returned callback twice. This effectively will lead either
to deadlock of the eventloop or earlier exit from a loop.

Given that it's never actually correct to that, panic seems okay.

We also make certain that if the user recovers from the panic we will
not have locked the eventloop either way,

In the future better APIs for most use cases should probably reduce the
need for people to use RegisterCallback and consequently worrying about
this.